add unit test framework

+++ SourceCode/Libraries/LibC/src/string.c
@@ -4,8 +4,8 @@
 
 #include <string.h>
 
-uint32_t strlen(const char *str) {
-  uint32_t len = 0;
+size_t strlen(const char *str) {
+  size_t len = 0;
   while (*str++ != '\0') {
     ++len;
   }
